+++ src/app/utils/CompanionApiUtils.js
@@ -305,6 +305,9 @@ function _addModifications(module, modifiers, blueprint, grade) {
     } else if (modifiers.modifiers[i].name === 'mod_weapon_burst_rof') {
       // For some reason this is a non-normalised percentage (i.e. 12.23% is 12.23 value rather than 0.1223 as everywhere else), so fix that here
       module.setModValue('burstrof', modifiers.modifiers[i].value * 100);
+    } else if (modifiers.modifiers[i].name === 'mod_weapon_falloffrange_from_range') {
+      // Obtain the falloff value directly from the range
+      module.setModValue('fallofffromrange', 1);
     } else {
       // Look up the modifiers to find what we need to do
       const modifierActions = Modifications.modifierActions[modifiers.modifiers[i].name];