+++ ChangeLog.md
@@ -1,5 +1,6 @@
 #2.2.9
   * Use SSL-enabled server for shortlinks
+  * Add falloff for weapons
 
 #2.2.8
   * Fix issue where filling all internals with cargo racks would include restricted slots